+++ 301/Exercises/LINQPracticeMinis/Solution/LINQPracticeMinis/Luchador.cs
@@ -7,11 +7,11 @@
         public string LastName { get; set; }
         public DateTime DoB { get; set; }
         public string Country { get; set; }
-        public int HeightCm { get; set; }
-        public int WeightKg { get; set; }
-        public int Wins { get; set; }
-        public int Losses { get; set; }
-        public int Draws { get; set; }
+        public decimal HeightCm { get; set; }
+        public decimal WeightKg { get; set; }
+        public decimal Wins { get; set; }
+        public decimal Losses { get; set; }
+        public decimal Draws { get; set; }
         public int Championships { get; set; }
     }
 }
